// src/app/shared/models/factura-recepcion.model.ts
// Generated by https://quicktype.io

import { Asociacion } from "./session-usuario.model";

export interface FacturaRecepcion {
    codigoAsociacion?: string;
    factura?:          Factura;
    sucursal?:         number;
    puntoVenta?:       number;
    usuario?:          string;
    asociacion:        Asociacion;
}

export interface Factura {
    cabecera: Cabecera;
    detalle:  Detalle[];
}

export interface Cabecera {
    codigoCliente?:       string;
    codigoExcepcion?:     number;
    codigoMetodoPago?:    number;
    codigoMoneda?:        number;
    montoTotal?:          number;
    montoTotalMoneda?:    number;
    descuentoAdicional?:  number;
    montoGiftCard?:       number;
    montoTotalSujetoIva?: number;
    numeroTarjeta?:      number;
    tipoCambio?:          number;
    idVenta?:             string|null;

    cafc?:                null;
    fechaEmision?:        string|null;
    numeroFactura?:       number;

    // otros campos adicionales para funcionaliodad
    cliente?:               any;
    nombreCliente?:        string|null;
    emailCliente?:        string|null;
    // campos clinica
    modalidadServicio?:        string|null;
    // campos alquiler
    periodoFacturado?:        string;
    // campos seguro
    ajusteAfectacionIva?:       number;
    // campos hotel
    cantidadHuespedes?:       number;
    cantidadHabitaciones?:       number;
    cantidadMayores?:       number;
    cantidadMenores?:       number;
    fechaIngresoHospedaje?:       string;
    // campos turismo
    razonSocialOperadorTurismo?:    string;
    // campos educativo
    estudiante?:               any;
    nombreEstudiante?:        string|null;
    idEstudiante?:        number|null;
    // ice
    montoIcePorcentual?:       number;
    montoIceEspecifico?:       number;
    // mineras
    datosEspecificos?: DatoEspecifico;
}

export interface Detalle {
    codigoProducto: string;
    producto?:      string;
    cantidad:       number;
    precioUnitario: number;
    monto:       number;
    montoDescuento: number;
    subTotal:       number;
    numeroImei?:     string;
    numeroSerie?:    string;
    // campos hospital
    especialidad?: string ;
	especialidadDetalle?: string;
	nroQuirofanoSalaOperaciones?: number;
	especialidadMedico?: string;
	nombreApellidoMedico?: string;
	nitDocumentoMedico?: number;
	nroMatriculaMedico?: string;
	nroFacturaMedico?: number;
    idTipoDescuento?: number;
    valorDescuento?: number;
    // campos hotel y turismo
    detalleHuespedes?: DetalleHuesped[];
    codigoTipoHabitacion?: number;
    tipoHabitacion?: string;
    descripcionAdicional?:     string;
    // ice
    marcaIce?: number;
    alicuotaIva?: number;
    precioNetoVentaIce?: number;
    alicuotaEspecifica?: number;
    alicuotaPorcentual?: number;
    montoIceEspecifico?: number;
    montoIcePorcentual?: number;
    cantidadIce?: number;
    datosEspecificos?: DetalleDatoEspecifico;
}

export interface DetalleHuesped {
    codigoPais: string;
    pais: string;
    documentoIdentificacion: string;
    nombreHuesped: string;
}

export interface DetalleDatoEspecifico {
    codigoNandina: string;
    descripcionLeyes: string;
    cantidadExtraccion: number;
    unidadMedidaExtraccion: number;
    unidadMedidaExtraccionDescripcion?: string;
}

export interface DatoEspecifico {
    direccionComprador?: string;
    concentradoGranel?: string;
    origen?: string;
    puertoTransito?: string;
    puertoDestino?: string;

    paisDestino?: number;
    incoterm?: string;
    tipoCambioAnb?: number;
    numeroLote?: string;
    kilosNetosHumedos?: number;
    humedadPorcentaje?: number;
    humedadValor?: number;
    mermaPorcentaje?: number;
    mermaValor?: number;
    kilosNetosSecos?: number;
    gastosRealizacion?: number;
    iva?: number;
    liquidacionPreliminar?: number;
    subTotal?: number;
    otrosDatos?: string;
    ruex?: string;
    nim?: string;

    tasaEfectiva?: number;
}